Check if unzipped path has one root folder
Check if the unzipped folder contains a single folder, which is the root
folder of the SIP. If there is no folder, more than one folder and/or one
or more files are present, the actual validation logic will not run and a
failed event will be sent out.

In case of success, pass the root folder of the SIP in the event instead
of the unzipped folder.

Furthermore, the validation logic got the zip folder as input. This is incorrect,
it should be the unzipped folder.

Also, run autoformat.